Learnt about @property tags. Also added sys.exit()
The chapter in Dave's book is talking over the finer details of implementing classes. Properties allow attributes to be accessed without calling on the bound method to return the value. Instead, the property is simply made when accessed. 

Added sys.exit() as a point of formality, not really required but in the future, this program will wake up, request a keyword, and upon a certain keypress, it will exit. 

Also, I have considered making the properties private, but since the information is access one-at-a-time instead of being part of an OO application, it would not fit into the format of the goal of this program; it would look odd to other programmers and even if it would create extremely robust code, for the user, it would not create the better experience - it would simply be backend cleanliness for the sake of cleanliness.